@@ -17,14 +17,14 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^19.1.7",
-    "@angular/common": "^19.1.7",
-    "@angular/compiler": "^19.1.7",
-    "@angular/core": "^19.1.7",
-    "@angular/forms": "^19.1.7",
-    "@angular/platform-browser": "^19.1.7",
-    "@angular/platform-browser-dynamic": "^19.1.7",
-    "@angular/router": "^19.1.7",
+    "@angular/animations": "^19.2.0",
+    "@angular/common": "^19.2.0",
+    "@angular/compiler": "^19.2.0",
+    "@angular/core": "^19.2.0",
+    "@angular/forms": "^19.2.0",
+    "@angular/platform-browser": "^19.2.0",
+    "@angular/platform-browser-dynamic": "^19.2.0",
+    "@angular/router": "^19.2.0",
     "@canvasjs/angular-charts": "^1.0.3",
     "bootstrap": "^5.3.1",
     "bootstrap-icons": "^1.9.1",
@@ -42,8 +42,8 @@
     "@angular-eslint/template-parser": "^18.0.1",
     "@typescript-eslint/parser": "^7.11.0",
     "@angular/cli": "^19.2.0",
-    "@angular/compiler-cli": "^19.1.7",
-    "@angular/language-service": "^19.1.7",
+    "@angular/compiler-cli": "^19.2.0",
+    "@angular/language-service": "^19.2.0",
     "@fortawesome/fontawesome-free": "^5.15.4",
     "@types/jasmine": "~3.6.0",
     "jasmine-core": "~3.8.0",
